[amux] Poll slave in a separate thread
When switching from one slave to another poll file descriptors change
and user is not mandatory to call snd_pcm_poll_descriptors() before each
time he calls poll(). The dup2() mechanism was here to transparently
change the poll fd.

Unfortunately some PCM does not only differ in poll file descriptor,
they also use different poll event. The dup2() mechanism counted on the
fact that listening to both POLLIN and POLLOUT event would do. But some
PCM plugin uses eventfd as a poll file descriptor which is always ready
on POLLOUT. Thus listening to both POLLIN and POLLOUT would unblock
poll() calls even if slave is not ready resulting (in the best case),
from the user point of view, of something similar to active polling.

To avoid that, now the slave polling is done in a separate thread and
user is notified through an eventfd.

Because creating a thread in a library is not a good option, the dup2()
behaviour can still be selected at compile time.