[amux] Fix the dup2 polling solution
The dup2 solution was let aside because polling on mocking fd for
POLLIN | POLLOUT events could result in non blocking poll.

The idea here is to separate the user pollfd array in two halves, one
half for POLLIN fd and the other half for POLLOUT fd. To poll a fd for
POLLIN | POLLOUT, this fd will be dup2'd in both halves.

Now the number of fd, their values and the event to poll for can change
without the user being notified.

The main pro here is that it avoids having to create a hidden thread
for every snd_pcm_open(). The cons is that it creates around 10 fakes fd
to work and is limited in the amount of pollfd a slave can have.

This is now the default polling mode.