poller/epoller: Add epoll based poller
Because epoll(7) FD is pollable, it can be used to multiplex polling on
a dynamic list of FD through a single one which makes it a suitable for
amux poller.

Sometimes the easiest solution takes time to stand out of the blue even
if it sits right under your nose.